Grid object has been removed, added intuitive gui
I thought adding the object Grid would simplify the way the canvas is updated, making it a single element that could be used in any canvas without the need of any external element. I think the way Paper.js handles the events makes this impossible. I changed the whole file trying to keep the code as simple as possible using global variables. The gui is functional even though it lacks the option of adding and removing columns and rows. Later a few beautifier elements will be added; some instructions and titles.